requirements.txt
streamlit
langchain-core
langchain-groq
langgraph
transformers
torch
chromadb
sentence-transformers
python-dotenv